GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Virtual branch (refs/gitbutler/Virtual-branch)
   branch head: 6f2f4d7ac247afa2e9ac3ce63d371b8e11a0247f
   - src/pages/index.tsx
   - .gitignore
   - package-lock.json
   - pnpm-lock.yaml

Your previous branch was: refs/heads/main

The sha for that commit was: 46bd15040b006078c621ce0f2a80631056794b91

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch